Start to rewrite deltas as component
