+++ web/index.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html class='h-full' lang='en'>
     <head>
-        <meta charset=utf-8 />
+        <meta charset='utf-8'/>
         <meta name='viewport' content='width=device-width, initial-scale=1'>
 
         <title>Hecate UI | Mapbox</title>
